[12065] Fix various warnings
Conflicts:
	src/game/BattleGroundAV.h
	src/game/CreatureLinkingMgr.cpp
	src/game/GridNotifiers.h
	src/game/LFGHandler.cpp
	src/game/Level3.cpp
	src/game/LootMgr.cpp
	src/game/MapPersistentStateMgr.cpp
	src/game/Object.h
	src/game/ObjectGridLoader.cpp
	src/game/ScriptMgr.cpp
	src/game/SpellAuras.cpp
	src/game/SpellAuras.h
	src/game/SpellMgr.cpp
	src/game/UnitAuraProcHandler.cpp
	src/game/WorldSession.h
	src/game/movement/MoveSpline.cpp
	src/game/movement/MoveSplineFlag.h

+++ src/game/ScriptMgr.cpp
@@ -44,7 +44,6 @@ INSTANTIATE_SINGLETON_1(ScriptMgr);
 
 ScriptMgr::ScriptMgr() :
     m_hScriptLib(NULL),
-    
     m_scheduledScripts(0),
 
     m_pOnInitScriptLibrary(NULL),
@@ -599,7 +598,7 @@ void ScriptMgr::LoadScripts(ScriptMapMapName& scripts, const char* tablename)
                         if (SpellEntry const* spell = sSpellStore.LookupEntry(i))
                             for (int j = 0; j < MAX_EFFECT_INDEX; ++j)
                             {
-                                if (spell->Effect[j] == SPELL_EFFECT_SEND_TAXI && spell->EffectMiscValue[j] == tmp.sendTaxiPath.taxiPathId)
+                                if (spell->Effect[j] == SPELL_EFFECT_SEND_TAXI && spell->EffectMiscValue[j] == int32(tmp.sendTaxiPath.taxiPathId))
                                 {
                                     taxiSpell = i;
                                     break;
